+++ jobs/Frontend/movie-search/src/hooks/__tests__/useSearchMovieQuery.test.tsx
@@ -1,46 +1,22 @@
 import { renderHook } from "@testing-library/react";
 import { useSearchMoviesQuery } from "../useSearchMoviesQuery";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { QueryClientProvider } from "@tanstack/react-query";
 import * as api from "../../api/requests";
 import { vi, describe, it, expect, beforeEach, afterEach } from "vitest";
 import { waitFor } from "@testing-library/react";
-
-function createClient() {
-  return new QueryClient({ defaultOptions: { queries: { retry: false } } });
-}
+import { createTestQueryClient } from "../../test-utils/wrappers";
+import { moviesPage1 } from "../../mocks/data";
 
 const wrapper = ({ children }: any) => (
-  <QueryClientProvider client={createClient()}>{children}</QueryClientProvider>
+  <QueryClientProvider client={createTestQueryClient()}>
+    {children}
+  </QueryClientProvider>
 );
 
 describe("useSearchMoviesQuery", () => {
-  const page1 = {
-    page: 1,
-    total_pages: 2,
-    total_results: 2,
-    results: [
-      {
-        id: 1,
-        title: "A",
-        adult: false,
-        backdrop_path: null,
-        genre_ids: [],
-        original_language: "en",
-        original_title: "A",
-        overview: "",
-        poster_path: null,
-        popularity: 0,
-        release_date: "2023-01-01",
-        video: false,
-        vote_average: 0,
-        vote_count: 0,
-      },
-    ],
-  };
-
   beforeEach(() => {
     vi.spyOn(api, "getSearchMovies").mockImplementation(() =>
-      Promise.resolve(page1)
+      Promise.resolve(moviesPage1)
     );
   });
 
@@ -53,7 +29,7 @@ describe("useSearchMoviesQuery", () => {
     expect(result.current.isLoading).toBe(true);
     await waitFor(() => result.current.isSuccess);
     await waitFor(() => {
-      expect(result.current.data?.pages[0]).toEqual(page1);
+      expect(result.current.data?.pages[0]).toEqual(moviesPage1);
     });
     expect(api.getSearchMovies).toHaveBeenCalledWith("foo", 1);
   });
